Update README.md
Hey Viggy, not sure what the final audience is for these reviews, but I thought we could add in some context in the README. I took a shot, but aimed to write it at a layperson's level. If this is for a more technical crowd, maybe we would not do as much about model and evaluation. Just wanted to throw out a PR!